requirements.txt
Django==1.9
Django-Select2==4.3.1
mysql-connector-python==2.0.4
PyMySQL==0.6.7
django-formset-js==0.5.0
django-tinymce==2.2.0
